useGetList for Getting Activity List
