+++ .expo/packager-info.json
@@ -2,9 +2,9 @@
   "devToolsPort": 19002,
   "expoServerPort": 19000,
   "packagerPort": 19003,
-  "packagerPid": 95733,
+  "packagerPid": 23066,
   "expoServerNgrokUrl": "https://xr-4p8.anonymous.mvp-map.exp.direct",
   "packagerNgrokUrl": "https://packager.xr-4p8.anonymous.mvp-map.exp.direct",
-  "ngrokPid": 95763,
+  "ngrokPid": 23094,
   "webpackServerPort": 19006
 }
